Hit the /finish-conversation endpoint on call end.
Hit the realtime /finish-conversation endpoint when the call ends, so
that we can find summaries by grepping for `TEXT_SUMMARY"` in the
COACHING directory of the run logs.

Postpone calling endSession() until we get the COACHING_END prompt
(previously, we closed it when we got the APP_TO_EXTERNAL_END_COACHING
message).

Accept sessionId in APP_TO_EXTERNAL_END_CALL, to be compatible with an
upcoming extension update that attempts to give better guarantees on
what happens if we get an END_CALL for the wrong call.

Of a sample of 2000 calls, 57 (2.85%) satisfied the
`lib.sell.long_enough_to_summarize()` heuristic.