Add "Click to request AI summary" button.
The button shows up if the coaching page receives a notification
that a text summary is not forthcoming. On clicking, the button
disappears and hits /request-text-summary on the services server.

The prompt ID, session ID, and client ID are all cached on the
button so that we can still record the Display of the summary---so
effectively the summary's Display record steals the prompt ID of the
TEXT_SUMMARY_PENDING=false message.

Switch `.copy-button:hover` to `button:hover` since the only two
preexisting button elements were of `copy-button` class.

Revive the "New!" callout and use `:after` so that we can apply it
to an element just by adding a css class.